+++ Modules/Inputs/Inputs/Data/InputControl.h
@@ -13,8 +13,8 @@ namespace Mani
 	struct AxisControl
 	{
 		std::string name = "NONE";
-		double x = 0.f;
-		double y = 0.f;
-		double z = 0.f;
+		float x = 0.f;
+		float y = 0.f;
+		float z = 0.f;
 	};
 }
